Skipping the test script test_ct_autoprovisionutils.py
```
TG-1060 - Skipping the test script test_ct_autoprovisionutils.py as this
script is failing and this needs to be rewritten.

Disabling this to make zuul to pass the ct.

This script is initially written to use lighttpd which is on switch, Now
lighttpd is not part of switch image is replaced by ops-restd.
Need to rewrite the script.

Tags: fix
```

Change-Id: I8e3fbfb0c6866731be5ffb0d749ebb718b09c645
Signed-off-by: Naresh Kumar <naresh.kumar8@hpe.com>